Changed readlink() function declaration and definition to match each other and POSIX.1-2001 standard.
git-svn-id: https://svn.aros.org/svn/aros/trunk@30231 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/unistd.h
@@ -252,7 +252,7 @@ int         ftruncate(int fd, off_t length);
 
 #if __POSIX_VISIBLE >= 200112 || __XSI_VISIBLE
 int         fchown(int fd, uid_t owner, gid_t group);
-ssize_t     readlink(const char * restrict path, char * restrict buf, int bufsize);
+ssize_t     readlink(const char * restrict path, char * restrict buf, size_t bufsize);
 #endif
 
 #if __POSIX_VISIBLE >= 200112
